DeploymentStoreClient: a fx dependency
