+++ proto/internal/temporal/server/api/deployment/v1/message.proto
@@ -34,7 +34,7 @@ import "temporal/api/common/v1/message.proto";
 message DeploymentLocalState {
     temporal.api.deployment.v1.Deployment worker_deployment = 1;
     map<string, TaskQueueFamilyInfo> task_queue_families = 2;
-    bool is_current_deployment = 3;
+    bool is_current = 3;
     google.protobuf.Timestamp create_time = 4;
     map<string, temporal.api.common.v1.Payload> metadata = 5;
