+++ service/matching/handler.go
@@ -43,6 +43,7 @@ import (
 	"go.temporal.io/server/common/persistence/visibility/manager"
 	"go.temporal.io/server/common/resource"
 	"go.temporal.io/server/common/tqid"
+	"go.temporal.io/server/service/worker/deployment"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
 )
@@ -78,6 +79,7 @@ func NewHandler(
 	taskManager persistence.TaskManager,
 	historyClient resource.HistoryClient,
 	matchingRawClient resource.MatchingRawClient,
+	deploymentStoreClient *deployment.DeploymentStoreClient,
 	hostInfoProvider membership.HostInfoProvider,
 	matchingServiceResolver membership.ServiceResolver,
 	metricsHandler metrics.Handler,
@@ -96,6 +98,7 @@ func NewHandler(
 			taskManager,
 			historyClient,
 			matchingRawClient, // Use non retry client inside matching
+			deploymentStoreClient,
 			config,
 			logger,
 			throttledLogger,
