+++ service/matching/matching_engine.go
@@ -71,6 +71,7 @@ import (
 	"go.temporal.io/server/common/tqid"
 	"go.temporal.io/server/common/util"
 	"go.temporal.io/server/common/worker_versioning"
+	"go.temporal.io/server/service/worker/deployment"
 	"google.golang.org/protobuf/types/known/durationpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
@@ -122,6 +123,7 @@ type (
 		taskManager                   persistence.TaskManager
 		historyClient                 resource.HistoryClient
 		matchingRawClient             resource.MatchingRawClient
+		deploymentStoreClient         *deployment.DeploymentStoreClient
 		tokenSerializer               common.TaskTokenSerializer
 		historySerializer             serialization.Serializer
 		logger                        log.Logger
@@ -188,6 +190,7 @@ func NewEngine(
 	taskManager persistence.TaskManager,
 	historyClient resource.HistoryClient,
 	matchingRawClient resource.MatchingRawClient,
+	deploymentStoreClient *deployment.DeploymentStoreClient,
 	config *Config,
 	logger log.Logger,
 	throttledLogger log.ThrottledLogger,
@@ -206,6 +209,7 @@ func NewEngine(
 		taskManager:                   taskManager,
 		historyClient:                 historyClient,
 		matchingRawClient:             matchingRawClient,
+		deploymentStoreClient:         deploymentStoreClient,
 		tokenSerializer:               common.NewProtoTaskTokenSerializer(),
 		historySerializer:             serialization.NewSerializer(),
 		logger:                        log.With(logger, tag.ComponentMatchingEngine),
