+++ service/worker/deployment/deployment_client.go
@@ -1,0 +1,198 @@
+// The MIT License
+//
+// Copyright (c) 2024 Temporal Technologies Inc.  All rights reserved.
+//
+// Copyright (c) 2024 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package deployment
+
+import (
+	"context"
+	"time"
+
+	commonpb "go.temporal.io/api/common/v1"
+	deploypb "go.temporal.io/api/deployment/v1"
+	enumspb "go.temporal.io/api/enums/v1"
+	taskqueuepb "go.temporal.io/api/taskqueue/v1"
+	updatepb "go.temporal.io/api/update/v1"
+	"go.temporal.io/api/workflowservice/v1"
+	deployspb "go.temporal.io/server/api/deployment/v1"
+	"go.temporal.io/server/api/historyservice/v1"
+	"go.temporal.io/server/common/namespace"
+	"go.temporal.io/server/common/payload"
+	"go.temporal.io/server/common/persistence/visibility/manager"
+	"go.temporal.io/server/common/primitives"
+	"go.temporal.io/server/common/sdk"
+	"go.temporal.io/server/common/searchattribute"
+	"google.golang.org/protobuf/types/known/timestamppb"
+)
+
+type DeploymentClient interface {
+	RegisterWorker(
+		ctx context.Context,
+		taskQueueName string,
+		taskQueueType enumspb.TaskQueueType,
+		pollTimestamp time.Time,
+	) error
+}
+
+// implements DeploymentClient
+type DeploymentStoreClient struct {
+	HistoryClient     historyservice.HistoryServiceClient
+	VisibilityManager manager.VisibilityManager
+}
+
+func (d *DeploymentStoreClient) RegisterTaskQueueWorker(
+	ctx context.Context,
+	namespaceEntry *namespace.Namespace,
+	deployment *deploypb.Deployment,
+	taskQueueName string,
+	taskQueueType enumspb.TaskQueueType,
+	pollTimestamp *timestamppb.Timestamp,
+	maxIDLengthLimit int,
+) error {
+	// validate params which are used for building workflowID's
+	err := ValidateDeploymentWfParams(SeriesFieldName, deployment.SeriesName, maxIDLengthLimit)
+	if err != nil {
+		return err
+	}
+	err = ValidateDeploymentWfParams(BuildIDFieldName, deployment.BuildId, maxIDLengthLimit)
+	if err != nil {
+		return err
+	}
+
+	deploymentWorkflowID := GenerateDeploymentWorkflowID(deployment.SeriesName, deployment.BuildId)
+	workflowInputPayloads, err := d.generateStartWorkflowPayload(namespaceEntry, deployment)
+	if err != nil {
+		return err
+	}
+	updatePayload, err := d.generateRegisterWorkerInDeploymentArgs(taskQueueName, taskQueueType, pollTimestamp)
+	if err != nil {
+		return err
+	}
+
+	sa := &commonpb.SearchAttributes{}
+	searchattribute.AddSearchAttribute(&sa, searchattribute.TemporalNamespaceDivision, payload.EncodeString(DeploymentNamespaceDivision))
+	searchattribute.AddSearchAttribute(&sa, searchattribute.DeploymentSeries, payload.EncodeString(deployment.SeriesName))
+
+	// initial memo fiels
+	memo, err := d.addInitialDeploymentMemo()
+	if err != nil {
+		return err
+	}
+
+	// Start workflow execution, if it hasn't already
+	startReq := &workflowservice.StartWorkflowExecutionRequest{
+		Namespace:                namespaceEntry.Name().String(),
+		WorkflowId:               deploymentWorkflowID,
+		WorkflowType:             &commonpb.WorkflowType{Name: DeploymentWorkflowType},
+		TaskQueue:                &taskqueuepb.TaskQueue{Name: primitives.PerNSWorkerTaskQueue},
+		Input:                    workflowInputPayloads,
+		WorkflowIdReusePolicy:    enumspb.WORKFLOW_ID_REUSE_POLICY_ALLOW_DUPLICATE,
+		WorkflowIdConflictPolicy: enumspb.WORKFLOW_ID_CONFLICT_POLICY_USE_EXISTING,
+		SearchAttributes:         sa,
+		Memo:                     memo,
+	}
+
+	updateReq := &workflowservice.UpdateWorkflowExecutionRequest{
+		Namespace: namespaceEntry.Name().String(),
+		WorkflowExecution: &commonpb.WorkflowExecution{
+			WorkflowId: deploymentWorkflowID,
+		},
+		Request: &updatepb.Request{
+			Input: &updatepb.Input{Name: RegisterWorkerInDeployment, Args: updatePayload},
+		},
+		WaitPolicy: &updatepb.WaitPolicy{LifecycleStage: enumspb.UPDATE_WORKFLOW_EXECUTION_LIFECYCLE_STAGE_COMPLETED},
+	}
+
+	// This is an atomic operation; if one operation fails, both will.
+	_, err = d.HistoryClient.ExecuteMultiOperation(ctx, &historyservice.ExecuteMultiOperationRequest{
+		NamespaceId: namespaceEntry.ID().String(),
+		WorkflowId:  deploymentWorkflowID,
+		Operations: []*historyservice.ExecuteMultiOperationRequest_Operation{
+			{
+				Operation: &historyservice.ExecuteMultiOperationRequest_Operation_StartWorkflow{
+					StartWorkflow: &historyservice.StartWorkflowExecutionRequest{
+						NamespaceId:  namespaceEntry.ID().String(),
+						StartRequest: startReq,
+					},
+				},
+			},
+			{
+				Operation: &historyservice.ExecuteMultiOperationRequest_Operation_UpdateWorkflow{
+					UpdateWorkflow: &historyservice.UpdateWorkflowExecutionRequest{
+						NamespaceId: namespaceEntry.ID().String(),
+						Request:     updateReq,
+					},
+				},
+			},
+		},
+	})
+
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// GenerateStartWorkflowPayload generates start workflow execution payload
+func (d *DeploymentStoreClient) generateStartWorkflowPayload(namespaceEntry *namespace.Namespace, deployment *deploypb.Deployment) (*commonpb.Payloads, error) {
+	workflowArgs := &deployspb.DeploymentWorkflowArgs{
+		NamespaceName: namespaceEntry.Name().String(),
+		NamespaceId:   namespaceEntry.ID().String(),
+		DeploymentLocalState: &deployspb.DeploymentLocalState{
+			WorkerDeployment:  deployment,
+			TaskQueueFamilies: nil,
+		},
+	}
+	return sdk.PreferProtoDataConverter.ToPayloads(workflowArgs)
+}
+
+// GenerateUpdateDeploymentPayload generates update workflow payload
+func (d *DeploymentStoreClient) generateRegisterWorkerInDeploymentArgs(taskQueueName string, taskQueueType enumspb.TaskQueueType,
+	pollTimestamp *timestamppb.Timestamp) (*commonpb.Payloads, error) {
+	updateArgs := &deployspb.RegisterWorkerInDeploymentArgs{
+		TaskQueueName:   taskQueueName,
+		TaskQueueType:   taskQueueType,
+		FirstPollerTime: nil, // TODO Shivam - come back to this
+	}
+	return sdk.PreferProtoDataConverter.ToPayloads(updateArgs)
+}
+
+func (d *DeploymentStoreClient) addInitialDeploymentMemo() (*commonpb.Memo, error) {
+	memo := &commonpb.Memo{}
+	memo.Fields = make(map[string]*commonpb.Payload)
+
+	deploymentWorkflowMemo := &deployspb.DeploymentWorkflowMemo{
+		CreateTime:          timestamppb.Now(),
+		IsCurrentDeployment: false,
+	}
+
+	memoPayload, err := sdk.PreferProtoDataConverter.ToPayload(deploymentWorkflowMemo)
+	if err != nil {
+		return nil, err
+	}
+
+	memo.Fields[DeploymentMemoField] = memoPayload
+	return memo, nil
+
+}
